+++ src/MultiDOF/Prog7/Multiforced/CFrameGraph.java
@@ -59,7 +59,7 @@ class CFrameGraph extends CFrame {
    }
 
    public void Frame(Graphics var1) {
-      byte var2 = 7;
+      int var2 = 7;
       boolean var3 = false;
       var1.setPaintMode();
       var1.setColor(Color.black);
@@ -76,7 +76,7 @@ class CFrameGraph extends CFrame {
       var1.setColor(Color.lightGray);
       var1.drawRect(super.x, super.y, super.width, super.height);
       var1.setColor(Color.black);
-      byte var6 = 3;
+      int var6 = 3;
       if (this.mMax == this.mMin) {
          var1.drawString("Please wait, calculating...", super.x + 10, super.y + 10);
       } else {
@@ -481,7 +481,7 @@ class CFrameGraph extends CFrame {
       }
 
       int var3 = 40;
-      byte var4 = 15;
+      int var4 = 15;
 
       for (int var1 = 1; var1 <= var2; var1++) {
          CSpringMassDamper var5 = this.GetSMD(var1);
